requirements.txt
streamlit
pandas
natsort
geopandas
folium
streamlit-folium
numpy
matplotlib
python-dotenv
sqlalchemy
psycopg2